added data from arduino

+++ Core/Inc/spi.h
@@ -48,9 +48,9 @@ void MX_SPI3_Init(void);
 void MX_SPI6_Init(void);
 
 /* USER CODE BEGIN Prototypes */
-void SPI_TransmitReceive(uint16_t* transferData, uint16_t* receiveData, uint16_t size);
-void SPI_Transfer(uint16_t* transferData, uint16_t size);
-void SPI_Receive(uint16_t* receiveData, uint16_t size);
+void SPI_TransmitReceive(uint8_t* transferData, uint8_t* receiveData, uint16_t size);
+void SPI_Transfer(uint8_t* transferData, uint16_t size);
+void SPI_Receive(uint8_t* receiveData, uint16_t size);
 void DMA1_ReceiveComplete(void);
 void DMA1_TransmitComplete(void);
 void SPI1_TransferError_Callback(void);
